#20 Refactor: 폴더 구조 및 패키지 정리
