+++ KAUPark/app/src/main/java/com/example/kaupark/PersonsAdapter.kt
@@ -4,6 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.kaupark.databinding.ListPersonBinding
+import com.example.kaupark.model.Person
 
 class PersonsAdapter(private val personList: MutableList<Person>) : RecyclerView.Adapter<PersonsAdapter.Holder>() {
