+++ KAUPark/app/src/main/java/com/example/kaupark/fragment/ChattingList.kt
@@ -1,10 +1,11 @@
-package com.example.kaupark
+package com.example.kaupark.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.kaupark.R
 
 
 class ChattingList : Fragment() {
