+++ KAUPark/app/src/main/java/com/example/kaupark/fragment/ParkingTimeLeft.kt
@@ -1,13 +1,13 @@
-package com.example.kaupark
+package com.example.kaupark.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.kaupark.databinding.HomeViewBinding
+import com.example.kaupark.databinding.ParkingTimeLeftBinding
 
-class HomeView : Fragment() {
+class ParkingTimeLeft : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -24,24 +24,15 @@ class HomeView : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val binding = HomeViewBinding.inflate(inflater, container, false)
+        val binding = ParkingTimeLeftBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     /*
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment HomeView.
-         */
-        // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            HomeView().apply {
+            ParkingTimeLeft().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
