+++ KAUPark/app/src/main/java/com/example/kaupark/model/Person.kt
@@ -1,4 +1,4 @@
-package com.example.kaupark
+package com.example.kaupark.model
 
 data class Person (val carNum : String,
                    val text : String,
