+++ KAUPark/app/src/main/res/layout/chatting_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChattingList" >
+    tools:context=".fragment.ChattingList" >
 
     <Button
         android:id="@+id/chatplus_btn"
