+++ KAUPark/app/src/main/res/layout/home_view.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeView">
+    tools:context=".fragment.HomeView">
 
         <androidx.appcompat.widget.Toolbar
             android:id = "@+id/hometoolbar"
