+++ KAUPark/app/src/main/res/layout/parking_time_left.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ParkingTimeLeft">
+    tools:context=".fragment.ParkingTimeLeft">
 
     <TextView
         android:id="@+id/txt_name"
